Mega commit of the chat function. It's not fully functional yet but it does send and receive messages which it stores in the database provided by firebase. Calling it a day and trying to fix it tomorrow before the presentation.

+++ app/src/main/java/com/example/datingapp/SecondProfileActivity.java
@@ -34,7 +34,7 @@ public class SecondProfileActivity extends AppCompatActivity {
     TextView textViewDisplayCity;
     TextView textViewDisplayDescription;
     TextView textViewDisplayQualifications;
-    int matchPercent;
+    Long matchPercent;
     TextView textViewMatchPercent;
 
     FirebaseAuth firebaseAuth = FirebaseAuth.getInstance();
@@ -190,7 +190,7 @@ public class SecondProfileActivity extends AppCompatActivity {
         userApi.matchUserAndCurrentProfile(signedInUserId, visitedProfileId).enqueue(new Callback<Integer>() {
             @Override
             public void onResponse(Call<Integer> call, Response<Integer> response) {
-                matchPercent = response.body();
+                matchPercent = Long.valueOf(response.body());
                 System.out.println("%: " + matchPercent);
                 textViewMatchPercent.setText("Ni matchar till " + matchPercent + "%");
             }
